Add a big-endian version of the thirdparty/icu conversion data and use it on s390x.
This allows unicode decoding to work correctly on big-endian platforms.
In particular, the tests:

* //tensorflow/python/kernel_tests/strings_ops:unicode_decode_op_test
* //tensorflow/python/kernel_tests/strings_ops:unicode_transcode_op_test

can now pass on big-endian platforms.

ICU version 69.1 was used to generate the big-endian data file to match the
latest little-endian thirdparty/icu update.

The instructions for generating the data files have been updated with the
steps to convert the little-endian files to big-endian so all files can
be generated at the same time on a little-endian machine.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>